+++ lib/shelly.rb
@@ -14,6 +14,7 @@ require "shelly/model"
 module Shelly
   autoload :App, "shelly/app"
   autoload :Cloudfile, "shelly/cloudfile"
+  autoload :Cloud, "shelly/cloud"
   autoload :Client, "shelly/client"
   autoload :StructureValidator, "shelly/structure_validator"
   autoload :User, "shelly/user"